- Now we detect when a schema has been partially applied. - Added option --check which will only check what has been applied and what will be applied. - Added option --noexecute which will prevent any db actions to be really performed.

+++ src/main/java/it/snova/dbmigrator/schema/SchemaMigrator.java
@@ -12,6 +12,7 @@ import java.io.IOException;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.util.ArrayList;
 import java.util.List;
 
 import org.apache.commons.io.IOUtils;
@@ -21,12 +22,16 @@ public class SchemaMigrator
   Connector connector;
   String encoding;
   boolean dropdb;
+  boolean checkscripts;
+  boolean noexecute;
   
   public SchemaMigrator(Connector connector)
   {
     this.connector = connector;
     this.encoding = "UTF-8";
     this.dropdb = false;
+    this.checkscripts = false;
+    this.noexecute = false;
   }
   
   public SchemaMigrator encoding(String encoding)
@@ -40,8 +45,74 @@ public class SchemaMigrator
     this.dropdb = dropdb;
     return this;
   }
+
+  public SchemaMigrator check(boolean check)
+  {
+    this.checkscripts = check;
+    return this;
+  }
+
+  public SchemaMigrator noexecute(boolean noexecute)
+  {
+    this.noexecute = noexecute;
+    return this;
+  }
   
   public int migrate(ScriptsCache scriptsCache)
+  {
+    if (checkscripts) {
+      return doCheckScripts(scriptsCache);
+    }
+    return doUpgrade(scriptsCache);
+  }
+
+  private int doCheckScripts(ScriptsCache scriptsCache)
+  {
+    try {
+      List<SchemaCache> schemas = scriptsCache.init();
+
+      int retvalue = 0;
+      
+      List<SchemaVersion> versions = getAllSchemaVersions();
+      for (SchemaVersion v: versions) {
+        System.out.println("check version " + v.toString());
+        
+        for (SchemaCache s: schemas) {
+          if (s.version().compareToIgnoreScript(v) == 0) {
+            for (String script: s.scripts()) {
+              File f = new File(script);
+              if (f.getName().equals(v.getScript())) {
+                String md5 = getMD5(f);
+                if (!md5.equals(v.getMD5())) {
+                  System.err.println("md5 of  " + script + " differs from applied!");
+                  retvalue = 1;
+                }
+                s.scripts().remove(script);
+                break;
+              }
+            }
+          }
+        }
+      }
+      for (SchemaCache s: schemas) {
+        List<String> scripts = s.scripts();
+        if (scripts.size() > 0) {
+          SchemaVersion v = s.version();
+          System.err.println("Missing " + scripts.size() + " script from version " + v.getMajor() + "." + v.getMinor() + "!");
+          for (String script: scripts) {            
+            System.err.println("Script: " + script);
+          }
+          retvalue = 1;
+        }
+      }
+      return retvalue;
+    } catch (IOException e) {
+      e.printStackTrace();
+    }
+    return 1;
+  }
+  
+  private int doUpgrade(ScriptsCache scriptsCache)
   {
     List<SchemaCache> schemas = scriptsCache.init();
     
@@ -50,12 +121,14 @@ public class SchemaMigrator
         recreateDB();
       }
       
+      List<SchemaVersion> versions = getAllSchemaVersions();
+
       SchemaVersion schema = getSchemaVersion();
       for (SchemaCache s: schemas) {
-        if (s.version().compareTo(schema) <= 0) {
+        if (s.version().compareTo(schema) < 0) {
           continue;
         }
-        applySchema(s);
+        applySchema(s, versions);
       }
       return 0;
     } catch (IOException e) {
@@ -65,24 +138,45 @@ public class SchemaMigrator
     }
     return 1;
   }
-  
-  private void applySchema(SchemaCache schema) throws IOException, SQLException
+
+  private void applySchema(SchemaCache schema, List<SchemaVersion> versions) throws IOException, SQLException
   {
-    System.out.println("Applying schema major: " + schema.version().getMajor() + " minor: " + schema.version().getMinor());
+    System.out.println("Applying schema major: " + schema.version().getMajor() + " minor: " + schema.version().getMinor() + "\n");
     
     Connect connect = connector.createConnect();
     
     for (String script: schema.scripts()) {
-      System.out.println("Script: " + script);
+      
+      File f = new File(script);
+      
+      boolean skip = false;
+      for (SchemaVersion v: versions) {
+        if (schema.version().compareToIgnoreScript(v) == 0) {
+          if (v.getScript().equals(f.getName())) {
+            String md5 = getMD5(f);
+            if (!md5.equals(v.getMD5())) {
+              System.err.println("the script " + script + " has already been imported but the md5 hash is not the same!");
+            }
+            skip = true;
+            break;
+          }
+        }
+      }
+      if (skip) {
+        continue;
+      }
+      System.out.println("* apply " + script + "\n");
 
       boolean versionExists = connect.doesTableExists("schemaversion");
       if (versionExists) {
         insertSchemaVersion(connect, schema.version(), script);
       }
       
-      ScriptRunner runner = new ScriptRunner(connect.getConnection(), false, true);
-      runner.setLogWriter(null);
-      runner.runScript(new FileReader(script));
+      if (!noexecute) {
+        ScriptRunner runner = new ScriptRunner(connect.getConnection(), false, true);
+        runner.setLogWriter(null);
+        runner.runScript(new FileReader(script));        
+      }
       
       if (!versionExists) {
         insertSchemaVersion(connect, schema.version(), script);
@@ -97,23 +191,30 @@ public class SchemaMigrator
     String content = IOUtils.toString(in, encoding);
     
     String description = getDescription(content);
-    System.out.println("Description: " + description);
+    System.out.println(description);
 
-    FileInputStream fis = new FileInputStream(f);
-    String md5 = org.apache.commons.codec.digest.DigestUtils.md5Hex(fis);
+    String md5 = getMD5(f);
     
-    Statement stmt = connect.createStatement();
-    String query = "insert into schemaversion "
-      + "(major, minor, date, script, description, md5) values "
-      + "(" + schema.getMajor()
-      + "," + schema.getMinor()
-      + ", NOW()"
-      + ", \"" + f.getName() + "\""
-      + ", \"" + description + "\""
-      + ", \"" + md5 + "\""
-      + ")";
-    stmt.executeUpdate(query);
-    stmt.close();    
+    if (!noexecute) {
+      Statement stmt = connect.createStatement();
+      String query = "insert into schemaversion "
+        + "(major, minor, date, script, description, md5) values "
+        + "(" + schema.getMajor()
+        + "," + schema.getMinor()
+        + ", NOW()"
+        + ", \"" + f.getName() + "\""
+        + ", \"" + description + "\""
+        + ", \"" + md5 + "\""
+        + ")";
+      stmt.executeUpdate(query);
+      stmt.close();
+    }
+  }
+  
+  private String getMD5(File f) throws IOException
+  {
+    FileInputStream fis = new FileInputStream(f);
+    return org.apache.commons.codec.digest.DigestUtils.md5Hex(fis);    
   }
   
   private String getDescription(String content)
@@ -131,24 +232,30 @@ public class SchemaMigrator
     return description.toString();
   }
 
-  private SchemaVersion getSchemaVersion()
+  private List<SchemaVersion> getAllSchemaVersions()
   {
-    SchemaVersion schema = new SchemaVersion(0, 0);
+    List<SchemaVersion> schemas = new ArrayList<SchemaVersion>();
+    
     try {
       Connect conn = connector.createConnect();
       if (conn.doesTableExists("schemaversion")) {
         Statement statement = conn.createStatement();
-        ResultSet resultSet = statement.executeQuery("select id, major, minor from schemaversion");
+        ResultSet resultSet = statement.executeQuery("select id, major, minor, script, description, md5 from schemaversion order by id");
         while (resultSet.next()) {
-          int id = resultSet.getInt(0);
-          int major = resultSet.getInt(1);
-          int minor = resultSet.getInt(2);
-          if (major > schema.getMajor()) {
-            schema.setMajor(major);
-            schema.setMinor(minor);
-          } else if (major == schema.getMajor() && minor > schema.getMinor()) {
-            schema.setMinor(minor);
-          }
+          int id = resultSet.getInt(1);
+          int major = resultSet.getInt(2);
+          int minor = resultSet.getInt(3);
+          String script = resultSet.getString(4);
+          String description = resultSet.getString(5);
+          String md5 = resultSet.getString(6);
+          
+          SchemaVersion schema = new SchemaVersion(major, minor);
+          schema.setScript(script);
+          schema.setDescription(description);
+          schema.setMD5(md5);
+          
+          schemas.add(schema);
+
           System.out.println("version : " + id + " major: " + major + " minor: " + minor);
         }
         statement.close();
@@ -156,16 +263,27 @@ public class SchemaMigrator
     } catch (SQLException e) {
       e.printStackTrace();
     }
-    return schema;
+    return schemas;
   }
-  
+
+  private SchemaVersion getSchemaVersion()
+  {
+    List<SchemaVersion> schema = getAllSchemaVersions();
+    if (schema.size() > 0) {
+      return schema.get(schema.size() - 1);
+    }
+    return new SchemaVersion(0, 0);
+  }
+
   private void recreateDB() throws SQLException
   {
-    Connect conn = connector.createConnect();
-    Statement statement = conn.createStatement();
-    statement.executeUpdate("drop schema " + connector.getDBName());
-    statement.executeUpdate("create schema " + connector.getDBName());
-    statement.close();
+    if (!noexecute) {
+      Connect conn = connector.createConnect();
+      Statement statement = conn.createStatement();
+      statement.executeUpdate("drop schema " + connector.getDBName());
+      statement.executeUpdate("create schema " + connector.getDBName());
+      statement.close();
+    }
   }
   
 }
